+++ CHANGELOG.md
@@ -5,19 +5,25 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-### Added
 ### Changed
-### Fixed
+- Moved logic related to the MultiChoiceRecyclerView class into the MultiChoiceAdapter
+- Internal improvements getting rid of holding views states
+- Aggregated some of the method of the MultiChoiceToolbar Builder like titles, defaultColors, selectedColors and change "setIcon" to "setDefaultIcon"
+
 ### Deleted
+- MultiChoiceRecyclerView class
 
 ## [1.2.14]
 ### Added
 - Possibility to disable some items from the selection mode with the method isSelectableInMultiChoiceMode(int position)
+
 ### Changed
 - Clean and polish
+
 ### Fixed
 - Kept order of selected items when retrieved back with "getSelectedItems"
 - Performance improvements
+
 ### Deleted
 - Application tag from library manifest
 - Removed unnecessary methods like setColumnNumber ad setRowNumber